Create environment slice
