+++ src/Core/Domain/EnvironmentDetailsModel.ts
@@ -18,12 +18,3 @@ export interface ModifyEnvironmentParams {
   icon?: string;
   description?: string;
 }
-
-export interface CreateEnvironmentParams {
-  name: string;
-  project_id: string;
-  repository?: string;
-  branch?: string;
-  icon?: string;
-  description?: string;
-}
