+++ src/Slices/CreateEnvironment/Core/CreateProjectCommand.ts
@@ -1,14 +1,14 @@
 import { ProjectModel } from "@/Core/Domain";
 import { Either } from "@/Core/Language";
 
-export interface CreateProject {
+export interface Command {
   kind: "CreateProject";
 }
-export interface CreateProjectManifest {
+export interface Manifest {
   error: string;
   apiData: string;
   body: { name: string };
-  command: CreateProject;
+  command: Command;
   trigger: (
     name: string
   ) => Promise<Either.Type<string, { data: ProjectModel }>>;
