+++ src/Slices/CreateEnvironment/Data/index.ts
@@ -1,0 +1,2 @@
+export * from "./CreateEnvironment";
+export * from "./CreateProject";
