+++ src/UI/Pages/PrimaryPageManager.tsx
@@ -5,6 +5,7 @@ import { AgentsPage } from "@S/Agents/UI";
 import { CompileDetailsPage } from "@S/CompileDetails/UI";
 import { CompileReportsPage } from "@S/CompileReports/UI";
 import { ComplianceCheckPage } from "@S/ComplianceCheck/UI";
+import { CreateEnvironmentPage } from "@S/CreateEnvironment/UI";
 import { CreateInstancePage } from "@S/CreateInstance/UI";
 import { DesiredStatePage } from "@S/DesiredState/UI";
 import { DesiredStateComparePage } from "@S/DesiredStateCompare/UI";
@@ -20,7 +21,6 @@ import { ResourcesPage } from "@S/Resource/UI/ResourcesPage";
 import { ResourceDetailsPage } from "@S/ResourceDetails/UI";
 import { ServiceInstanceHistoryPage } from "@S/ServiceInstanceHistory/UI";
 import { SettingsPage } from "@S/Settings/UI";
-import { CreateEnvironmentPage } from "./CreateEnvironment";
 import { HomePage } from "./Home";
 import { ServiceCatalogPage } from "./ServiceCatalog";
 import { ServiceInventoryPage } from "./ServiceInventory";
